Misc tweaks to DDG graph node
