+++ packages/jaeger-ui/src/components/DeepDependencyGraph/getNodeLabel/index.tsx
@@ -1,4 +1,4 @@
-// Copyright (c) 2017 Uber Technologies, Inc.
+// Copyright (c) 2019 Uber Technologies, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,4 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default } from './getNodeLabel';
+import * as React from 'react';
+
+import DdgNode from './DdgNode';
+import { DdgVertex, StyleStates } from '../../../model/ddg/types';
+
+// temp fill in props
+/* istanbul ignore next */
+const noops = {
+  setViewModifier(vertexKey: string, viewModifier: StyleStates, enabled: boolean) {
+    // eslint-disable-next-line no-console
+    console.log(`set view modifier: ${enabled ? 'on' : 'OFF'} ${viewModifier} -- ${vertexKey}`);
+  },
+};
+
+export default function getNodeLabel(vertex: DdgVertex) {
+  // temp -- get DdgVertex for rendering purposes
+  const pathElem = [...vertex.pathElems][0];
+  if (!pathElem) {
+    throw new Error('Invalid vertex state');
+  }
+  const { name: operationName, service } = pathElem.operation;
+  const isFocalNode = pathElem.memberIdx === pathElem.memberOf.focalIdx;
+  return (
+    <DdgNode
+      vertexKey={vertex.key}
+      service={service.name}
+      operation={operationName}
+      isFocalNode={isFocalNode}
+      viewModifiers={0}
+      focalNodeUrl={isFocalNode ? null : 'some-url'}
+      {...noops}
+    />
+  );
+}
